requirements.txt
fastapi[standard]==0.115.4
pydantic==2.9.2
black==24.8.0
uvicorn==0.31.0
alembic==1.14.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
psycopg2-binary==2.9.10
sqlmodel==0.0.22
python-jose==3.3.0
cffi==1.17.1
cryptography==44.0.0
pycparser==2.22
loguru==0.6.0
pymongo==4.10.1
pika==1.3.2
jsonpickle==4.0.1
iniconfig==2.0.0
pluggy==1.5.0
pytest==8.3.4
pytest-asyncio==0.25.0